refactor(reviews): collapse aggregates into one file with internal hydrate/list split
Single service function with one Zod input schema (union of hydrate +
paginated). Auth and rubric setup happen once at the entry, then dispatch
to runHydrate / runList by input shape. Router calls the one entry point
without dispatch logic of its own.

Drops the shared and hydrate files; the with-block is a plain helper
function and buildItem is generic so types infer from the findMany call
sites rather than from a type-only probe.
